Report: Initial version of long-waiting commits
Commits that are changed a long time after a previous version. The
versions may occur at any point in the history of the file, but are
contiguous (although deleted and recreated files with the same name are
counted as one). For now, we have some limitations to ensure the query
does not collect too many possible matches.