Sprint definitions: Only use complete date if earlier than end date
The close date should be the date when the team is done with the works
and are still within time. This makes the visualisations more in line
with expectations when we have both complete date and end date;
otherwise sprints would overlap even more because some stories are not
set to be completed correctly.