Reports: Add option to use project IDs instead of names
This makes the data more anonymous and safer to use by not directly
showing identifiable names of projects. Note that the timeline still has
files that contain the project name from the query results, and that
some reports may contain information that allows one to deduce the
project (e.g. changed paths), so not everything is viable for demos.