Export source types as localization
This allows direct use by the JSON output for localization via the
visualization-ui component.

Handle separately from other localization data (features etc.) since the
YAML format is different and the icons need to be treated differently
from a multi-column feature, a principle that does not exist here (so
lists must be preserved and nested).