Features: Correct sprint number using conditions
When we filter the sprints using conditions, then the sprint number
feature would incorrectly count the filtered sprints (but not add them
to the output). This causes mismatches between sprint metadata that is
not available during prediction and the sprint number and must thus be
avoided. Recursive interpolation now allows selecting a table to apply
the conditions on.