Expected story points: More late planning leeway
Allow adding stories to the sprint up to one day from the registered
start of the sprint, because this is sometimes a time outside of working
hours and the actual sprint start would occur when the stories are added
the day after. This is simply to ensure we get valid planned points, and
is separate from the other early/late allowed plan/work intervals.