Planned story points: Exclude altered/removed case
If the story points are changed before removing the story from the
sprint, then we would incorrectly use the old story points and not
consider the sprint change. This is now corrected for closed sprints,
where ongoing sprints can still make use of the story points while the
sprint is ongoing.