Build: Drop default image, only provide dashboard image
The default image only contains the dependencies which is not enough for
our purposes of collecting data for specific visualizations, predictions
or analysis reports. The dashboard image is also the only viable image
for running in a standalone manner (BigBoat). The default image has no
purpose and is thus no longer built.