Sprint features: Velocity expression
Replace the query for 1-sprint velocity (points/day) with an expression
that provides the done story points per weekday.

The weekday feature is added, which calculates the days within the
sprint which are working days using 7/5 calculation with adjustment for
starting a sprint on a later weekday than it ends. This gives the most
sprints in our dataset a weekday count divisible by 5 (10 or 15).