Features: Migrate metrics to separate queries
Only collect metrics that are defined in the features and are selected
in the collection arguments. This means it is no longer possible to
collect all metrics from the database and we instead curate the metrics
we wish to have. This has initial support for different aggregate
functions (min/max/avg/sum) and adds metadata to the feature definition
to place it in the correct category and provide accurate logging.