Combine concurrent sprints of same project
Initial work on combining sprints that take place at the same time.
Currently only one concurrency metric is used, e.g., start or end date.
This merges all features, correcting the handling of expression metrics
to adjust a window or recalculate their merged value on the new dataset.
Some metadata is merged (notably not the sprint name and close date).
Details tables are also merged.

This does not handle future sprints well yet and is only done if the
--combine argument is provided with a metadata field of the sprint.