Recent sprint features: Improve query performance
Do not perform 2*P queries where P is the number of projects in the
database when we wish to include both old and new sprints. Instead use
one query in this case and after all data is collected, update the
recentness of the sprints. This speeds up the queries significantly.