Features: Replace team projects with board filter
Individual projects that have a board specified for them should be
filtered to only contain the sprints for that board as well as the other
conditions (start/end date), such that the projects provide an accurate
display of the team's uses of that project.

The new projects are given team IDs but are not considered teams.

Remove a sort on sprint name which is not necessary for combined sprints
and which does not work once the sprint name contains some lists.