Sources: Remove unboxing of source IDs
This keeps the output format of projects with a single domain name the
same as that of a project with multiple domain names, because the
unboxing isn't clever enough to work for both types. Keeping a JSON
array here is fine and lets us extend this format with more source IDs
at a later point.