Use more variables in sprint definitions/features
All sprint definitions must now be expanded using the recursive string
interpolation instead of used directly. Additionally, we have functions
to interpolate a sequence of field names into columns with a table name
or into join columns of two table names. The field names may be provided
from a definition or from a contextual variable such as the join columns
used to combine the features of the same sprint.

This makes it more viable to use the same features on tables from
a different source or to adjust the join columns to group sprints in
a different way than on project.