Details for components and teams with components
Due to improper inheritance of project IDs and component metadata of
these projects, both the component and the team were considered to be
a component, amd the project IDs were set to exclude the original ID of
the project containing the components. By avoiding this the details for
teams with components and components within teams is reinstated.