Features: Component can remove details from others
Features with details that are collected for a project with components
can now be filtered such that an issue does not count toward both
a component with a 'remove' key and another (or no) component. This
avoids double-counting in some cases and allows filtering an entire
component from the result set (i.e. by marking the component non-main).

This only works if the details have a key; the summarization is done
afterward. Convert a few story features to count via summarization.