Features: Do not combine metrics across sprints of same project
Most metrics are already combined using the details, but if this is not
possible (e.g. technical debt with carry-over or wrapped coverage), then
by default use the latest sprint's value if we combine sprints of the
same project. The behavior for teams with multiple projects is not
changed and the coverage actually performs its own combining for now.

This avoids over-counting technical debt when there are (partially)
overlapping sprints.