Features: Readjust original component on (non)full projects
If a component is provided from an inclusion/exclusion filter then we
should take appropriate action to keep the original component, when the
component field already existed. This is because the rows may be
duplicated when the project seems to consist of only one component, but
the initial query must provide all valid components so that feature
queries can join to them correctly.