Do not consider patch sprints for backlog interval sprints
This avoids miscounting stories into a sprint that is not used in the
result, by providing the list of useful sprints to the queries that look
for changes to stories made during a sprint. Additionally this amounts
to a performance increase to the query (from 10+ seconds to 1 second).

Compiling queries also uses sprint conditions now.