Feature to count modified points on backlog relative to average points
This feature counts the difference points compared to what we would fill
in for the backlog estimation, if there was no older points change. This
brings this feature closer to what our estimation considers and is thus
viable for addition in the simulation with all backlog mutations, where
modified points were missing previously.