Weather: Use new KNMI API, config in main file
Use new API with key to be set in configuration file in order to
retrieve the URL to the measurements file. By using configuration, the
API key stored in a credentials file not part of the repository, and
additionally one can set different configuration per organization,
different lat/lon values for example.