Sprint results: Include date ranges of dataset
Specify the minimum and maximum dates of sprints in the configuration of
the model, and update the combined configuration based on the sprints of
all organizations if multiple were involved.

This may be useful to specify (e.g. in JSON-LD) the temporal coverage of
the dataset that the model was trained on. This also ensures likeness of
the configuration files across multiple involved organizations.