Update Dockerfile to latest R version
Because CRAN packages keep updating, it is difficult for us to keep the
dependencies in a static state. Especially when the package necessary to
install libraries at a specific version (devtools) itself depends on
packages that are 4+ only now. Also, the debian image does not contain
the correct keys to pull from APT repositories, which requires insecure
overrides. So building new Docker images or other development
environments with version 3.6.x is very tedious.

Most analysis scripts have not been tested for R version 4 and up, but
for now we support local development with the Dockerfile.